requirements.txt
streamlit
pandas
altair==4.0
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.8.2
Pillow>=9.1.0